refactor: 피드백을 적용합니다.
- Line Column 어노테이션 추가
- PathService 중복 로직 제거
- Line 조회 쿼리 최적화
- infra 영역에 해당하지 않는 클래스들에 대한 패키지 이동
