+++ src/main/java/nextstep/subway/path/domain/AgeDiscount.java
@@ -1,8 +1,7 @@
-package nextstep.subway.path.infra;
+package nextstep.subway.path.domain;
 
 import nextstep.subway.auth.domain.LoginMember;
 import nextstep.subway.common.Money;
-import nextstep.subway.path.domain.MemberDiscount;
 import org.springframework.stereotype.Component;
 
 import java.util.Arrays;
